Refactoring of some methods around for a cleaner app
