+++ social_relay/utils/data.py
@@ -1,18 +1,32 @@
 # -*- coding: utf-8 -*-
+import json
+
 import redis
-from rq import Queue, Worker
 
 from social_relay import config
 
 
 r = redis.Redis(host=config.REDIS_HOST, port=config.REDIS_PORT, db=config.REDIS_DB)
 
-public_queue = Queue("receive", connection=r)
-
 
 def get_pod_preferences():
     return r.hgetall("pod_preferences")
 
 
-def get_worker_count():
-    return len(Worker.all(public_queue.connection))
+def nodes_who_want_tags(tags):
+    nodes = []
+    for node, data in get_pod_preferences().items():
+        data = json.loads(data.decode("utf-8"))
+        if not set(data["tags"]).isdisjoint(tags):
+            # One or more tags match
+            nodes.append(node.decode("utf-8"))
+    return nodes
+
+
+def nodes_who_want_all():
+    nodes = []
+    for node, data in get_pod_preferences().items():
+        data = json.loads(data.decode("utf-8"))
+        if data["subscribe"] and data["scope"] == "all":
+            nodes.append(node.decode("utf-8"))
+    return nodes
